python/src/otel/otel_sdk/requirements.txt
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-distro==0.47b0
opentelemetry-instrumentation==0.47b0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-propagator-aws-xray==1.0.1

# Instrumentation dependencies
opentelemetry-instrumentation-aiohttp-client==0.47b0
opentelemetry-util-http==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-asyncpg==0.47b0
opentelemetry-instrumentation-boto==0.47b0
opentelemetry-instrumentation-boto3sqs==0.47b0
opentelemetry-instrumentation-celery==0.47b0
opentelemetry-instrumentation-dbapi==0.47b0
opentelemetry-instrumentation-django==0.47b0
opentelemetry-instrumentation-elasticsearch==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-falcon==0.47b0
opentelemetry-instrumentation-flask==0.47b0
opentelemetry-instrumentation-grpc==0.47b0
opentelemetry-instrumentation-jinja2==0.47b0
opentelemetry-instrumentation-mysql==0.47b0
opentelemetry-instrumentation-psycopg2==0.47b0
opentelemetry-instrumentation-pymemcache==0.47b0
opentelemetry-instrumentation-pymongo==0.47b0
opentelemetry-instrumentation-pymysql==0.47b0
opentelemetry-instrumentation-pyramid==0.47b0
opentelemetry-instrumentation-redis==0.47b0
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-instrumentation-sqlalchemy==0.47b0
opentelemetry-instrumentation-sqlite3==0.47b0
opentelemetry-instrumentation-starlette==0.47b0
opentelemetry-instrumentation-tornado==0.47b0
opentelemetry-instrumentation-wsgi==0.47b0
